refactor - Refactored the Margin Implementation to the Label Implementation
